See if breaking up lsl.misc.ionosphere makes it easier to maintain.
